requirements.txt
asyncpg==0.29.0
mcp==0.4.0
python-dotenv==1.0.0
requests==2.32.4